Move items from Security to Auth namespace
