+++ DLCS.Repository.Tests/Strategy/BasicHttpAuthOriginStrategyTests.cs
@@ -4,8 +4,8 @@ using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Threading.Tasks;
 using DLCS.Core.Types;
+using DLCS.Model.Auth;
 using DLCS.Model.Customers;
-using DLCS.Model.Security;
 using DLCS.Repository.Strategy;
 using FakeItEasy;
 using FluentAssertions;
