+++ DLCS.Repository/Auth/DapperCredentialsRepository.cs
@@ -2,18 +2,17 @@
 using System.IO;
 using System.Threading.Tasks;
 using DLCS.Core.Guard;
+using DLCS.Model.Auth;
 using DLCS.Model.Customers;
-using DLCS.Model.Security;
 using DLCS.Model.Storage;
 using DLCS.Repository.Caching;
-using DLCS.Repository.Settings;
 using LazyCache;
 using Microsoft.Extensions.Caching.Memory;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Newtonsoft.Json;
 
-namespace DLCS.Repository.Security
+namespace DLCS.Repository.Auth
 {
     /// <summary>
     /// Implementation of <see cref="ICredentialsRepository"/> using Dapper for data access
@@ -48,7 +47,7 @@ namespace DLCS.Repository.Security
 
             return appCache.GetOrAddAsync(cacheKey, () =>
             {
-                logger.LogDebug("Refreshing CustomerOriginStrategy credentials for {customerOriginStrategy}",
+                logger.LogDebug("Refreshing CustomerOriginStrategy credentials for {CustomerOriginStrategy}",
                     customerOriginStrategy.Id);
                 return GetBasicCredentials(credentials, customerOriginStrategy.Id);
             }, cacheSettings.GetMemoryCacheOptions(priority: CacheItemPriority.Low));
@@ -70,18 +69,21 @@ namespace DLCS.Repository.Security
             }
             catch (Exception ex)
             {
-                logger.LogError(ex, "Error getting credentials for customerOriginStrategy {customerOriginStrategy}", id);
+                logger.LogError(ex, "Error getting credentials for customerOriginStrategy {CustomerOriginStrategy}", id);
                 throw;
             }
         }
         
         private static bool CredentialsAreForS3(string credentials) => credentials.StartsWith("s3://");
 
-        private async Task<BasicCredentials> GetBasicCredentialsFromBucket(string credentials)
+        private async Task<BasicCredentials?> GetBasicCredentialsFromBucket(string credentials)
         {
             var objectInBucket = RegionalisedObjectInBucket.Parse(credentials);
             var credentialStream = await bucketReader.GetObjectContentFromBucket(objectInBucket);
-            using var reader = new StreamReader(credentialStream);
+            
+            if ((credentialStream ?? Stream.Null) == Stream.Null) return null;
+            
+            using var reader = new StreamReader(credentialStream!);
             using var jsonReader = new JsonTextReader(reader);
             return jsonSerializer.Deserialize<BasicCredentials>(jsonReader);
         }
