+++ DLCS.Repository/DlcsContext.cs
@@ -5,10 +5,10 @@ using DLCS.Core.Enum;
 using DLCS.Model.Assets;
 using DLCS.Model.Assets.CustomHeaders;
 using DLCS.Model.Assets.NamedQueries;
+using DLCS.Model.Auth;
 using DLCS.Model.Customers;
-using DLCS.Model.Security;
+using DLCS.Repository.Auth;
 using DLCS.Repository.Entities;
-using DLCS.Repository.Security;
 using DLCS.Repository.Serialisation;
 using Microsoft.EntityFrameworkCore;
 using Newtonsoft.Json;
