+++ Orchestrator.Tests/Integration/FileHandlingTests.cs
@@ -3,8 +3,8 @@ using System.Net.Http;
 using System.Threading.Tasks;
 using Amazon.S3;
 using DLCS.Model.Assets;
+using DLCS.Model.Auth;
 using DLCS.Model.Customers;
-using DLCS.Model.Security;
 using FluentAssertions;
 using LazyCache;
 using LazyCache.Mocks;
