+++ Orchestrator.Tests/Integration/ImageHandlingTests.cs
@@ -10,7 +10,7 @@ using Amazon.S3;
 using Amazon.S3.Model;
 using DLCS.Core.Collections;
 using DLCS.Core.Types;
-using DLCS.Model.Security;
+using DLCS.Model.Auth;
 using DLCS.Repository.Entities;
 using FluentAssertions;
 using Microsoft.AspNetCore.Mvc.Testing;
