+++ Orchestrator/Features/Images/Requests/GetImageInfoJson.cs
@@ -5,7 +5,7 @@ using DLCS.Core;
 using DLCS.Core.Collections;
 using DLCS.Core.Types;
 using DLCS.Model.Assets;
-using DLCS.Model.Security;
+using DLCS.Model.Auth;
 using DLCS.Web.Requests.AssetDelivery;
 using DLCS.Web.Response;
 using IIIF.Serialisation;
