+++ Orchestrator/Infrastructure/ServiceCollectionX.cs
@@ -4,15 +4,15 @@ using API.Client;
 using DLCS.Core.Encryption;
 using DLCS.Model.Assets;
 using DLCS.Model.Assets.CustomHeaders;
+using DLCS.Model.Auth;
 using DLCS.Model.Customers;
 using DLCS.Model.PathElements;
-using DLCS.Model.Security;
 using DLCS.Repository;
 using DLCS.Repository.Assets;
 using DLCS.Repository.Assets.CustomHeaders;
+using DLCS.Repository.Auth;
 using DLCS.Repository.Caching;
 using DLCS.Repository.Customers;
-using DLCS.Repository.Security;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
